swatches workaround
Swatches are seemingly impossible to implement with the current setup (no sign of already implemented swatches in the website, no access to products to try and add swatches myself).

Workaround involves piece of code (marked with comment) that retrieves the names of the colors, which then link to the correct variant.